Move functions related to exercises into its own file
- Renamed some variables
- sweet.go handles command flags and exercise parsing
- exercise.go handles the bubbletea application (model and rendering)
- moved shared code into util package
